feat: useFormAuth
